adjust utils
